Removal of js and addition of php
Transfer this back into a PHP app. On original version, Javascript was
handling the displaying of images on the index page which makes no
sense and only done to prove a point.  It has been transferred back to
PHP and JS is there to move carousel.  Also, until a better solution is
found, the login system is reverted back to the original version.